+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `EmbeddedRenderer` to create a `mp4`-video which needs ffmpeg installed and findable via path, for more info see the `Dockerfile` ([#64])
 - Added `FrameRenderer` to store the individual frames on the disk ([#90])
 - Added `RemoteRenderer` to render composition on remote server ([#94])
+- Added feature `save_load` to save and load an composition from disk ([#94])
 - Implemented animation `ChangeColorAnimation` ([#64])
 - Implemented animation `PositionAnimation` ([#64])
 - Implemented animation `SetColorAnimation` ([#64])